pyproject.toml
[tool.poetry]
name = "curvy-ai"
version = "0.1.0"
description = "A bot that plays curve fever"
authors = ["Douglas Li <douglas@shiftlab.ai>"]

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.6.0"
numpy = "^1.23.2"
mss = "^6.1.0"
pynput = "^1.7.6"
tesserocr = "^2.5.2"
Pillow = "^9.2.0"
torch = "^1.12.1"
torchvision = "^0.13.1"
torchaudio = "^0.12.1"
gym = {extras = ["classic_control"], version = "^0.25.2"}
matplotlib = "^3.5.3"
pyppeteer = "^1.0.2"
termcolor = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
